requirements.txt
torch
transformers
pillow
pytesseract
easyocr
opencv-python
safetensors
google-colab
